chore(access-tests-use-dummy): Access tests Now Use `dummy`.
For the time being, events will not be available in production. Events
will not be an immediately necessary feature so it can wait.

Added :func:`resolve_model` to :mod:`models`.
Added command ``client config docker-mysql``.
Refactoring of ``test_access``. Added parameterizations.
Added ``resolvable_model`` arguments to :meth:`get_primary` of
:class:`Dummy`. Added ``own`` argument to derivative functions using
callbacks. Added randomization when empty results are not turned on to
minimize the number of exceptions arising from :class:`Dummy` (it
works!).