feature(auth): Added Demo Accounts and Minimal UI. Token Uses ``sub`` (RFC rfc7519-4.1).
See https://www.rfc-editor.org/rfc/rfc7519#section-4.1, global
uniqueness of the 'subject' field applies here. Removed the ``uuid`` field
from ``Token`` as it would be inconvenient to implement with auth0.
Somehow this information would have to be added to auth0 therefore requiring
use of the API. Instead, the ``sub`` field is used to uniquely identify any
incoming users and is saved as a shasum in the database.
Updated dockerfile reduce number of ENV layers.
Added templates.
Added ``/register``, ``/logout``, ``/profile``, and ``/demo``.